FAQ_Chatbot/requirements.txt
langchain
langchain-community
langchain-core
langchain-openrouter
sentence-transformers
pypdf
faiss-cpu
streamlit
python-dotenv
streamlit-lottie
requests